Reworked CFG construction to allow tags both in states and on transitions.
This is needed in preparation for implementing staDFA algorithm, in which
tag operations are placed in states, not on transitions. The new CFG model
is able to accommodate both TDFA and staDFA, as well as a hypothetical
"mixed" case where tag operations occur both in states and on transitions.

This commit adds "canonical" marking of visited states and transitions, which
allows to avoid erroneous creation of dulicate arcs in the CFG.

Some of the helper functions have been rewritten from recursion to iteration.

This commit also adds CFG construction context to avoid passing too many
parameters to the helper functions.