Don't allocate unused tag variable.
In some cases this was causing increase in the number of tag variables after
optimizations, which is generally bad and also caused memory access violation
as the liveness/interference tables are allocated for the initial number.

The 0-th variable is superficial and does not correspond to a real tag version
(zero is a special version reserved for "no tag operation"). In the variable
allocation algorithm it is convenient to handle zero in the same way as other
variables (because arrays are indexed from zero). In most of the cases 0-th
variable ends up in a class that contains other (non-fake) variables, but if it
is alone in its class, then no tag variable should be allocated for this class.