Handle block-local liveness in dead code elimination.
This is needed in preparation for staDFA implementation. Variables local to
basic block may occur as a result of breaking cycles in a sequence of tag
commands by introducing temporary variables. TDFA mappings have no cycles
(such mappings are rejected), but staDFA transfer actions often do, because
the indices used in the actions are bound to TNFA substates, which may be
flipped in the origin and target staDFA states.