Consistently remove duplicates when optimizing in tag commands.
Duplicate commands may arise as a result of tag optimizations, and should be
removed in all cases ("set', "copy" and "add" commands). In case of "set"
commands this can be done by sorting commands and removing adjacent duplicates.
In case of "add" commands sorting is not possible, but deduplication can still
be implemented using naive quadratic algorithm. Deduplication of "copy" commands
depends on the presence of cycles (if there are no cycles, it is possible to
sort commands lexicographically, then remove adjecent duplicates, then sort
commands topologically).