Added internal option '--posix-prectable <naive | complex>'.
The new option allows to specify the algorithm used to compute POSIX precedence
table. The 'complex' algorithm computes precedence table in one traversal of
tag history tree and has quadratic complexity in the number of TNFA states; it
is the default. The 'naive' algorithm has worst-case cubic complexity in the
number of TNFA states, but it is much simpler than 'complex' and may be slightly
faster in non-pathological cases.