Rewrite recursion into iteration (skeleton max path length computation).
Ensured that the new algorithm computes the same length as the old one
by adding temporary debug print of the computed length to both
algorithms and verifying that it coincides.

This is to avoid stack overflow on large RE (especially on instrumented
builds that have larger stack frames, like AddressSanitizer).

Stack overflow reported by Agostino Sarubbo.