Avoid using complex expressions in function arguments.
Although the current behavior was ok (only one of the arguments was
a complex expression, and it didn't rely on side-effects), such code
patterns are error-prone and may cause subtle bugs in the future.

Note:

Re2c uses C++98 standard. C++11 adds some guarantees in section 1.9.15:
(http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2011/n3242.pdf):

  Every evaluation in the calling function (including other function
  calls) that is not otherwise specifically sequenced before or after
  the execution of the body of the called function is indeterminately
  sequenced with respect to the execution of the called function.

And a footnote:

  In other words, function executions do not interleave with each other.

But it is does not solve the problem of possible side-effects.