Go backend: change default options (set generic API, unset decorations).
It is necessary to have different sets of default mutable options for
different language backends. For example, all Go programs need to set
generic API, so it makes sense to set it by default.

The difficulty in changing default mutable options based on the constant
(global) options is that re2c allows to intermix mutable and constant
options on the command line. Therefore by the time all constant options
have been parsed, some of the mutable options may have already been set.
It is impossible to tell which of them have been set from the option
values, because a user-set value may coincide with the default value.
Therefore this commit adds boolean flags indicating if a mutable option
has been explicitly set.