Align default value of 're2c:*:naked' options with 're2c:decorate'.
Now instead of always defaulting to false, 're2c:*:naked' options
default to the value of 're2c:decorate' (unless they are explicitly set
by the user). The idea behind this change is to allow the user set one
option instead of a dozen, while still allowing a more fine-grained
control if necessary.