Clean up goto elision. Apply it more aggressively with EOF rule.
Previously goto elision was implemented by passing a null target state
instead of the actual DFA state to the code generation function. However
that prevented goto elision in some other cases with EOF rule, because
the decision whether to elide or not was based on the target state
(which was force-set to null).

This patch adds an explicit "elide" flag instead of the null "hint".