Don't reserve usused initial label for skeleton programs.
This was a weird special case in order to preserve code generation
results during the major rework of the code generation subsystem. It is
no longer needed, as the changes in tests are trivial, concern only a
few skeleton tests and can be checked manually (they are caused by a
shift in label numbers).