Do not add hoisted tags to fallback transitions.
A precondition to hoisting tag operations out of conditional branches
is that all transitions agree on tags, including the potential fallback
(final or default) transition. Therefore tag operations should not be
repeated on fallback transition. With s-tags repeating is redundant but
harmless, but with m-tags it could cause incorrect submatch results.