Do not accumulate options from rules/reuse blocks in whole-program options.
This is necessary for consistency with YYFILL states (`-f` option),
which are block-local in reuse mode (`-r` option).

Re2c only recently started to allow mixing reuse and normal blocks,
and the primary user was re2c own lexer, so this change is unlikely
to break anyone.

This partially fixes #282 "Options -f, -c and -r do not combine well".