Decouple start label and initial label.
Start label is needed in `-f` mode: it points to state 0 (the beginning
of block, before condition dispatch in `-c` mode). Initial label points
to the beginning of the DFA (after condition dispatch in `-c` mode).
Previously re2c reused the same label in cases when both start and
initial labels were needed and could be placed adjacent to each other.
Aliasing made it difficult to change the code, because changes of the
start label affected the initial label (and caused massive diffs in the
tests due to a shift in the label numbers).

This patch decouples the labels: now start label, if needed, has its own
label, and initial label always has it own label.

There are 29 changed tests. The changes can be grouped in 3 categories:

1. Trivial renumbering of labels. In most cases all labels except `yy0`
   (which is the start label) have been shifted by 1. It is possible to
   mechanically ensure that nothing else has changed using this script:

    for f in `find . -name '*.orig'`; do
        sed -r 's/(.*)(yy)([1-9][0-9]*)(.*)/cat << EOF\n\1\2$((\3 + 1))\4\nEOF/ge' $f > $f.mod
    done

   In most cases the ".mod" file (which is generated from the original
   test result by incrementing all labels with sed) is identical to the
   new tests result. In some cases manual verfication is necessary
   because labels shifted by a varying number (e.g. due to multiple
   reuse blocks in one file: shift was 2 for the first block, 4 for the
   second block, etc.).

2. Flip of adjacent user-defined start label and re2c-defined start
   label. This happened because previously re2c-defined start label
   aliased to the initial label (which is emitted after the user-defined
   start label), and re2c generated the the initial label (the start
   label was hidden). Now there is no alias, and start label is no
   longer hidden. The change is harmless because user-defined start
   label and re2c-defined start label are always generated next to each
   other.

3. Added jump from the start label to the initial label. Previosly the
   state dispatch jumped directly to the initial label, and now it jumps
   to the start label, and from there to the initial label. The extra
   jump should be trivial to optimise for the compiler.