Get rid of useless label increment.
For historic reasons in `-c` mode re2c started label enumeration from 1
rather than 0. Label 0 is reserved for the start state with `-f` option,
or if the start label is explicitly set with `re2c:startlabel`
configuration. In other cases there is no reason to skip index 0 (except
to avoid changes in the generated code). This patch makes a small change
in the code and drops the artificial increment.

The changes in tests are trivial: label numbers have been decreased
by 1 (or by 2, 4, etc. in tests with multiple reuse blocks). The
following perl script decrements all labels in the code generated by
re2c before the change (old.c) and compares the result (old1.c) with the
new code (new.c). The expected diff is empty. This doesn't work for all
tests (e.g. reuse mode), but in most cases it is enough to verify that
nothing else has changed:

    perl -pe 's/(yy|YYDEBUG\()(\d+)/($1).($2-1)/eg' old.c > old1.c
    diff old1.c new.c