Add named placeholders in interpolated configuration strings.
Previously re2c allowed to use @@ or a custom string as a placeholder
for an argument in API primitives. With the addition of multi-argument
APIs like YYSHIFTSTAG and YYSHIFTMTAG this has become ambiguous. The
first attempt to resolve ambiguity was to use @@1, @@2, etc. This works,
but it is not perfect for two reasons. First, it is not clear how to
handle multi-digit suffixes: @@12 can be @1 followed by symbol 2, or it
can be @12 (but it is unlikely that re2c will ever have and API with 12
arguments, so it is most likely an error). Second, it is hard to
remember what is the meaning of each numbered placeholder.

This commit adds named placeholders of the form @@{name}. For APIs that
have only one argument short form @@ is still recognized (if the longer
form failed), but for multi-argument APIs only named form is allowed.

It is possible to globally override @@ with any string, for example to
have placeholders of the form ${name} or #{name}.