Fix GCC-10 warning -Wreturn-local-addr.
Previously the function was not erroneous, because only the head of the
newly constructed list was in a local stack variable, and the rest
consisted of reordered elements from the input list. The offending
pointer was initially set to the head of the list (local addr), and
then advanced to subsequent elements. The case when it was not advanced
and still pointed at the local variable was handled in a special way by
returning NULL.

This patch slightly rewrites topsort to use the output parameter
instead of a local variable as the head of the new topsorted list.