Make filenames in configurations relative to the source file directory.
Prepend source file directory to these filenames when parsing
configurations and updating the corresponding options.

The affected configurations are:
    `re2c:flags:o`, `re2c:flags:output`
    `re2c:flags:t`, `re2c:flags:type-header`

The reason for this change is to make the contents of the source files
independent of the subdirectory where they are located relative to the
build directory where re2c is run. Otherwise it is necessary to hardcode
the relative path in the configuration and change it every time when
either the source file is moved to another subdirectory, or the build
directory is changed.

This change breaks backwards compatibility. However, the configurations
have been added not very long ago (in re2c-1.2) and not documented yet,
so it is likely that not many users will be affected. The corresponding
command-line options `-o`, `--output` and `-t`, `--type-header` remain
unchanged.