run_tests.sh: do not copy each test into temporary subdirectory before running.
Previously we did this to avoid tracking all the files generated by a
test run (optional header files, skeleton artifacts, etc.) However, this
approach has problems with tests that generate header in a nested
subdirectory: directory structure needs to exist (otherwise re2c won't
be able to open the header file), but re-creating it in a temporary
subdirectory is problematic (`find` collides with other threads running
in parallel, creating and removing directories).

Instead of using a sandbox, track all files generated by a test (that
turns out to be not so difficult with shell globs).

This patch adds a requirement that all tests that use header must
specify it explicitly in the test options (in filename) using the long
option --type-header. Also, this patch allows using plus sign in the
argument to --type-header option to substitute for slash (that cannot be
used in filenames). This allows one to specify headers in nested
subdirectories.