Remove `re2c:flags:output` and make `re2c:flags:type-header` relative to the output directory.
Before this patch the filename specified in `re2c:flags:t`,
`re2c:flags:type-header` configuration was relative to the directory
of the source file. It is inconvenient for out-of-tree builds, when the
output is generated in a directory different from the source directory
(which may be read-only). Therefore this patch makes the filename in
configuration relative to the directory of the output file.

As for the `re2c:flags:o`, `re2c:flags:output` configuration, it is
removed altogether. Making it relative to the source directory has the
same problems as with the header file, and making it absolute or
relative to the build directory (where re2c is run) is also bad,
because build-related information will have to be hard-coded in the
source file. Given that this is a relatively new configuraion and it
hasn't been documented in any released version of re2c, it is better to
delete it.

The change does not affect `-t`, `--type-header` and `-o`, `--output`
command-line options.