Do not reset API-related configurations in each `use:re2c` block.
This change breaks backwards compatibility for some (rare) use cases.

For historical reasons re2c reset all API-related configurations to
default values in each `/*!use:re2c ... */` block. This behaviour is
illogical and inconvenient: the whole purpose of reuse mode is to share
configurations, definitions and rules between blocks, and resetting
selected configurations means that they have to be redefined in each
reuse block. This has caused confusion (see bug #291 "Defines in rules
block are not propagated to use blocks" reported by Joshua Inscoe).

The odd behavior was kept for backwards compatibility. It goes back to
the versions of re2c and was probably introduced accidentally or by
mistake. It is unlikely that many users rely on the old behavior: if
they need user-defined configurations in some of the blocks and default
configurations in other blocks, it is more likely that they would
explicitly reset them to defaults rather than rely on undocumented re2c
behavior.

The change affects the following configurations:

    re2c:define:YYGETCONDITION
    re2c:define:YYSETCONDITION
    re2c:define:YYFILL
    re2c:define:YYGETSTATE
    re2c:define:YYSETSTATE
    re2c:define:YYBACKUP
    re2c:define:YYBACKUPCTX
    re2c:define:YYCONDTYPE
    re2c:define:YYCTXMARKER
    re2c:define:YYCTYPE
    re2c:define:YYCURSOR
    re2c:define:YYDEBUG
    re2c:define:YYLESSTNAN
    re2c:define:YYLIMIT
    re2c:define:YYMARKER
    re2c:define:YYPEEK
    re2c:define:YYRESTORE
    re2c:define:YYRESTORECTX
    re2c:define:YYRESTORETAG
    re2c:define:YYSTAGN
    re2c:define:YYSTAGP
    re2c:define:YYMTAGN
    re2c:define:YYMTAGP
    re2c:define:YYSHIFT
    re2c:define:YYSHIFTSTAG
    re2c:define:YYSHIFTMTAG
    re2c:define:YYSKIP
    re2c:label:yyFillLabel
    re2c:label:yyNext
    re2c:variable:yyaccept
    re2c:variable:yybm
    re2c:variable:yych
    re2c:variable:yyctable
    re2c:variable:yytarget