Avoid ambiguity in the generated DFA names in the presence of include files.
If the definition of a re2c block comes from an include file, use the name of
the file as well as the block line to generate a unique DFA name. The block
line itself is insufficient as there may be two blocks on the same line in
different files.

This has caused an error in one of the skeleton tests.