Add fixed-tag optimization for nested tags (part 1).
When applying fixed tags optimization to tags that are enclosed in
alternative or repetition it is necessary to check if the base tag has a
no-match value (in that case fixed tag should also be set to no-match).
For tags in top-level concatenation the check is not needed, because
they always match.

This commit implements the optimization, but does not enable it yet
(that would require updatimg many test results and will be done in a
separate commit).

This is related to bug #281 "Apply fixed tags optimization more
aggressively".