Remove duplicate tests for POSIX capturing groups.
The tests are generated with a Bash script from a list of test cases
that contain a regexp and a string. Some test cases use the same regexp
multiple times with a different string. For the autogenerated tests the
string is not used, so such test cases result in identical tests.

This patch changes the script to check for duplicates. The original set
of test cases remains unchanged and can be run with another script:

    test/posix_captures/.run/__run.sh