Fix example for in-tree builds.
This example uses include file unicode_categories.re and specifies
paths to include directory using -I option. Since the test can be
run both in in-tree and out-of-tree builds, different include paths
are needed to cover different cases.

In addition it is necessary to disable line directives, otherwise the
unstable path to the include file gets embedded in the output file.