Submatch benchmarks: add "ab" series.
These are artificial benchmarks for sequences of 'a' and 'b' symbols
with varying degrees of tag nondeterminism. The regular expression is:

    R(n,k) = @a_1 a* @b_1 a{k} b* ... @a_n a* @b_n a{k} b*

For example, R(2,4) = @a_1 a* @b_1 aaaa b* @a_2 a* @b_2 aaaa b*. The
benchmark includes cases R(2,0), R(2,4), R(2,8), R(4,0), R(4,2), R(4,4),
R(8,0), R(8,1), R(8,2).

Some of the benchmarks in this series cannot be handled correctly by
Ragel due to its lack of support for nondeterministic tags.

Kleenex size and sta-DFA performance degrade quickly with increasing
level of nondeterminism (the k-parameter).