Add option `--depfile <filename>` to generate build dependency file.
The new option is similar to GCC or Clang option `-MD -MF <filename>`.
It computes a transitive closure of all files included with
`/*!include:re2c ... */` directives and writes them to the specified
file in the form of a makefile rule:

    <output> : <input1> <input2> ...

The new option requires that the output file is specified (otherwise it
is unclear how to generate dependency rule).